// apps/game-client/src/lib/api/generated/main/model/create-manual-timer-dto-type.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */

export type CreateManualTimerDtoType = typeof CreateManualTimerDtoType[keyof typeof CreateManualTimerDtoType];


export const CreateManualTimerDtoType = {
  ELITE2: 'ELITE2',
  ELITE3: 'ELITE3',
  HERO: 'HERO',
  TITAN: 'TITAN',
} as const;